+++ poms/transactions/serializers.py
@@ -44,6 +44,8 @@ from poms.transactions.models import TransactionClass, Transaction, TransactionT
     TransactionTypeActionInstrumentEventScheduleAction
 from poms.users.fields import MasterUserField
 
+from poms.common.utils import date_now
+
 
 class EventClassSerializer(PomsClassSerializer):
     class Meta(PomsClassSerializer.Meta):
@@ -1529,8 +1531,8 @@ class ComplexTransactionMixin:
         return data
 
 
-class ComplexTransactionSerializer(ComplexTransactionMixin, ModelWithAttributesSerializer):
-    text = serializers.SerializerMethodField()
+class ComplexTransactionSerializer(ModelWithAttributesSerializer):
+    # text = serializers.SerializerMethodField()
     master_user = MasterUserField()
     transaction_type = serializers.PrimaryKeyRelatedField(read_only=True)
     transactions = serializers.PrimaryKeyRelatedField(read_only=True, many=True)
@@ -1550,6 +1552,54 @@ class ComplexTransactionSerializer(ComplexTransactionMixin, ModelWithAttributesS
             'id', 'date', 'status', 'code', 'text', 'is_deleted', 'transaction_type', 'transactions', 'master_user'
         ]
 
+    def update(self, instance, validated_data):
+
+        data = super(ComplexTransactionSerializer, self).to_representation(instance)
+
+        transaction_type_process_instance = TransactionTypeProcess(transaction_type=instance.transaction_type,
+                                                                   complex_transaction=instance,
+                                                                   context=self.context)
+
+        if instance.transaction_type.display_expr:
+            ctrn = formula.value_prepare(data)
+            trns = ctrn.get('transactions', None)
+
+            names = {
+                'complex_transaction': ctrn,
+                'transactions': trns,
+            }
+
+            for key, value in transaction_type_process_instance.values.items():
+                names[key] = value
+
+            try:
+                validated_data['text'] = formula.safe_eval(instance.transaction_type.display_expr, names=names,
+                                                           context=self.context)
+            except formula.InvalidExpression:
+                validated_data['text'] = '<InvalidExpression>'
+
+        if instance.transaction_type.date_expr:
+            ctrn = formula.value_prepare(data)
+            trns = ctrn.get('transactions', None)
+
+            names = {
+                'complex_transaction': ctrn,
+                'transactions': trns,
+            }
+
+            for key, value in transaction_type_process_instance.values.items():
+                names[key] = value
+
+            try:
+                validated_data['date'] = formula.safe_eval(instance.transaction_type.date_expr, names=names,
+                                                           context=self.context)
+            except formula.InvalidExpression:
+                validated_data['date'] = date_now()
+
+        instance = super(ComplexTransactionSerializer, self).update(instance, validated_data)
+
+        return instance
+
 
 class ComplexTransactionEvalSerializer(ComplexTransactionSerializer):
     def __init__(self, *args, **kwargs):
@@ -1558,7 +1608,7 @@ class ComplexTransactionEvalSerializer(ComplexTransactionSerializer):
 
 
 class ComplexTransactionViewSerializer(ComplexTransactionMixin, serializers.ModelSerializer):
-    text = serializers.SerializerMethodField()
+    # text = serializers.SerializerMethodField()
     transaction_type = serializers.PrimaryKeyRelatedField(read_only=True)
 
     class Meta: